requirements.txt
mysql-connector-python==8.1.0
requests
python-dateutil
python-dateutil
passlib
uvicorn
fastapi
itsdangerous
keyring
cryptography
python-multipart
gunicorn
werkzeug
pillow
fernet
eyed3
asyncio
websockets
pytz
psycopg2-binary
mygpoclient
feedparser
pyotp
appdirs
argon2_cffi
httpx
qrcode[pil]
pyjwt